requirements_new.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.11.0
decorator==4.4.2
einops==0.7.0
executing==1.2.0
fairscale==0.4.13
filelock @ file:///work/ci_py311/filelock_1676823024085/work
fire==0.5.0
fonttools==4.42.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.9.2
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
html-testRunner==1.2.1
huggingface-hub==0.18.0
hydra-core==1.3.2
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.31.5
imageio-ffmpeg==0.4.9
ipython==8.15.0
jedi==0.19.0
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
libretranslatepy==2.1.1
lit==17.0.1
lxml==4.9.3
Markdown==3.5
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
meshcat==0.3.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
moviepy==1.0.3
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
natsort==8.4.0
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp311-cp311-linux_x86_64.whl#sha256=184178a893ff0b5e0cb1b0b7a0dddff49dc91424c4e4a01e29bcb05f27a698a6
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.1
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1696580024257/work
pointnet2-ops @ file:///home/duanj1/m2t2/manipulate_anything/pointnet2_ops
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyglet==2.0.9
Pygments==2.16.1
pyngrok==7.0.0
PyOpenGL==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyquaternion==0.9.9
pyrender==0.1.45
PyRep @ file:///home/duanj1/m2t2/manipulate_anything/PyRep
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rlbench @ file:///home/duanj1/m2t2/manipulate_anything/RLBench
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentencepiece==0.1.99
six==1.16.0
stack-data==0.6.2
sympy @ file:///work/ci_py311_2/sympy_1679339311852/work
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
timeout-decorator==0.5.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.31.0
transformers-stream-generator==0.0.4
translate==3.6.1
trimesh==3.23.5
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
u-msgpack-python==2.8.0
urllib3==2.0.6
wcwidth==0.2.6
Werkzeug==3.0.0
yarl==1.9.2
yarr @ file:///home/duanj1/m2t2/manipulate_anything/YARR
